fixup! fix: Enable more linters
