+++ .golangci.yml
@@ -1,4 +1,4 @@
-# Documentation reference https://github.com/golangci/golangci-lint/blob/v1.55.2/.golangci.reference.yml
+# Documentation reference https://github.com/golangci/golangci-lint/blob/v1.56.2/.golangci.reference.yml
 run:
   skip-dirs-use-default: false
   modules-download-mode: readonly
@@ -61,6 +61,13 @@ linters-settings:
     rules:
       - name: line-length-limit
         disabled: true
+      - name: function-length
+        disabled: true
+      # TODO remove
+      - name: cyclomatic
+        disabled: true
+      - name: cognitive-complexity
+        disabled: true
   unparam:
     check-exported: true
 
@@ -114,9 +121,6 @@ issues:
     - linters:
         - revive
       text: "^add-constant: avoid magic numbers like '1', create a named constant for it$"
-    # - linters:
-    #     - stylecheck
-    #   text: "ST1000:|ST1020:|ST1021:|ST1022:"
   max-issues-per-linter: 0
   max-same-issues: 0
