+++ cmd/main.go
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+// Entrypoint of the project
 package main
 
 import (
