+++ internal/handler/velerobackup_handler.go
@@ -14,7 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+// Package handler contains all event handlers of the project
+package handler
 
 import (
 	"context"
@@ -25,9 +26,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	"github.com/migtools/oadp-non-admin/internal/common/constant"
 )
 
-// VeleroBackupHandler Handles VeleroBackup events
+// VeleroBackupHandler contains event handlers for Velero Backup objects
 type VeleroBackupHandler struct {
 	Logger logr.Logger
 }
@@ -36,6 +39,7 @@ func getVeleroBackupHandlerLogger(ctx context.Context, name, namespace string) l
 	return log.FromContext(ctx).WithValues("VeleroBackupHandler", types.NamespacedName{Name: name, Namespace: namespace})
 }
 
+// Create event handler
 func (*VeleroBackupHandler) Create(ctx context.Context, evt event.CreateEvent, _ workqueue.RateLimitingInterface) {
 	nameSpace := evt.Object.GetNamespace()
 	name := evt.Object.GetName()
@@ -43,6 +47,7 @@ func (*VeleroBackupHandler) Create(ctx context.Context, evt event.CreateEvent, _
 	logger.V(1).Info("Received Create VeleroBackupHandler")
 }
 
+// Update event handler
 func (*VeleroBackupHandler) Update(ctx context.Context, evt event.UpdateEvent, q workqueue.RateLimitingInterface) {
 	nameSpace := evt.ObjectNew.GetNamespace()
 	name := evt.ObjectNew.GetName()
@@ -56,13 +61,13 @@ func (*VeleroBackupHandler) Update(ctx context.Context, evt event.UpdateEvent, q
 		return
 	}
 
-	nabOriginNamespace, ok := annotations[NabOriginNamespaceAnnotation]
+	nabOriginNamespace, ok := annotations[constant.NabOriginNamespaceAnnotation]
 	if !ok {
 		logger.V(1).Info("Backup NonAdminBackup origin namespace annotation not found")
 		return
 	}
 
-	nabOriginName, ok := annotations[NabOriginNameAnnotation]
+	nabOriginName, ok := annotations[constant.NabOriginNameAnnotation]
 	if !ok {
 		logger.V(1).Info("Backup NonAdminBackup origin name annotation not found")
 		return
@@ -74,10 +79,12 @@ func (*VeleroBackupHandler) Update(ctx context.Context, evt event.UpdateEvent, q
 	}})
 }
 
+// Delete event handler
 func (*VeleroBackupHandler) Delete(_ context.Context, _ event.DeleteEvent, _ workqueue.RateLimitingInterface) {
 	// Delete event handler for the Backup object
 }
 
+// Generic event handler
 func (*VeleroBackupHandler) Generic(_ context.Context, _ event.GenericEvent, _ workqueue.RateLimitingInterface) {
 	// Generic event handler for the Backup object
 }
