+++ internal/predicate/composite_predicate.go
@@ -14,7 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+// Package predicate contains all event filters of the project
+package predicate
 
 import (
 	"context"
@@ -22,12 +23,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/event"
 )
 
+// CompositePredicate is a combination of all project event filters
 type CompositePredicate struct {
 	Context                 context.Context
 	NonAdminBackupPredicate NonAdminBackupPredicate
 	VeleroBackupPredicate   VeleroBackupPredicate
 }
 
+// Create event filter
 func (p CompositePredicate) Create(evt event.CreateEvent) bool {
 	// If NonAdminBackupPredicate returns true, ignore VeleroBackupPredicate
 	if p.NonAdminBackupPredicate.Create(p.Context, evt) {
@@ -37,6 +40,7 @@ func (p CompositePredicate) Create(evt event.CreateEvent) bool {
 	return p.VeleroBackupPredicate.Create(p.Context, evt)
 }
 
+// Update event filter
 func (p CompositePredicate) Update(evt event.UpdateEvent) bool {
 	// If NonAdminBackupPredicate returns true, ignore VeleroBackupPredicate
 	if p.NonAdminBackupPredicate.Update(p.Context, evt) {
@@ -46,6 +50,7 @@ func (p CompositePredicate) Update(evt event.UpdateEvent) bool {
 	return p.VeleroBackupPredicate.Update(p.Context, evt)
 }
 
+// Delete event filter
 func (p CompositePredicate) Delete(evt event.DeleteEvent) bool {
 	// If NonAdminBackupPredicate returns true, ignore VeleroBackupPredicate
 	if p.NonAdminBackupPredicate.Delete(p.Context, evt) {
@@ -55,6 +60,7 @@ func (p CompositePredicate) Delete(evt event.DeleteEvent) bool {
 	return p.VeleroBackupPredicate.Delete(p.Context, evt)
 }
 
+// Generic event filter
 func (p CompositePredicate) Generic(evt event.GenericEvent) bool {
 	// If NonAdminBackupPredicate returns true, ignore VeleroBackupPredicate
 	if p.NonAdminBackupPredicate.Generic(p.Context, evt) {
