+++ internal/predicate/velerobackup_predicate.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package predicate
 
 import (
 	"context"
@@ -24,8 +24,11 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/migtools/oadp-non-admin/internal/common/function"
 )
 
+// VeleroBackupPredicate contains event filters for Velero Backup objects
 type VeleroBackupPredicate struct {
 	// We are watching only Velero Backup objects within
 	// namespace where OADP is.
@@ -33,10 +36,13 @@ type VeleroBackupPredicate struct {
 	Logger              logr.Logger
 }
 
+// TODO try to remove calls to get logger functions, try to initialize it
+
 func getBackupPredicateLogger(ctx context.Context, name, namespace string) logr.Logger {
 	return log.FromContext(ctx).WithValues("VeleroBackupPredicate", types.NamespacedName{Name: name, Namespace: namespace})
 }
 
+// Create event filter
 func (veleroBackupPredicate VeleroBackupPredicate) Create(ctx context.Context, evt event.CreateEvent) bool {
 	nameSpace := evt.Object.GetNamespace()
 	if nameSpace != veleroBackupPredicate.OadpVeleroNamespace {
@@ -52,12 +58,13 @@ func (veleroBackupPredicate VeleroBackupPredicate) Create(ctx context.Context, e
 		// The event object is not a Backup, ignore it
 		return false
 	}
-	if HasRequiredLabel(backup) {
+	if function.CheckVeleroBackupLabels(backup) {
 		return true
 	}
 	return false
 }
 
+// Update event filter
 func (veleroBackupPredicate VeleroBackupPredicate) Update(ctx context.Context, evt event.UpdateEvent) bool {
 	nameSpace := evt.ObjectNew.GetNamespace()
 	name := evt.ObjectNew.GetName()
@@ -67,10 +74,12 @@ func (veleroBackupPredicate VeleroBackupPredicate) Update(ctx context.Context, e
 	return nameSpace == veleroBackupPredicate.OadpVeleroNamespace
 }
 
+// Delete event filter
 func (VeleroBackupPredicate) Delete(_ context.Context, _ event.DeleteEvent) bool {
 	return false
 }
 
+// Generic event filter
 func (VeleroBackupPredicate) Generic(_ context.Context, _ event.GenericEvent) bool {
 	return false
 }
